Try to use VC UUID first for locating VMs, then fall back to BIOS UUID
When communication from VMCI vSocket is established n vmdk_ops.py, we
find out VM ID from
the socket, and then locate VM ManagedObject by this ID.

There are 2 IDs - VC UUID and BIOS UUID. When .vmx file is
copied (e.g. VCD or other products), BIOS UUID can be duplicate thus
failures like #1373 .

However, VC UUID is unique on creation of VM, and unique as long as ESXi
is a
part of a VC. So it is a much better candidate for using as a unique ID.

This change tries to use VC UUID first, and fails back to BIOS UUID
next. The change also checks that the VM name (for a found VM by ID)
matches the one we see from vSocket, and does a few other minor prints.